Change slot position to a short to keep object size below 32 bytes

+++ rhino/src/main/java/org/mozilla/javascript/Slot.java
@@ -18,7 +18,7 @@ public class Slot implements Serializable {
     Object value;
     transient Slot next; // next in hash table bucket for EmbeddedSlotMap only
     transient Slot orderedNext; // next in linked list for EmbeddedSlotMap only
-    transient int orderedPos; // Position in the array for OrderedSlotMap only
+    transient short orderedPos; // Position in the array for OrderedSlotMap only
 
     Slot(Object name, int index, int attributes) {
         this.name = name;
